# Python: Optimize PyArrow reads