@Fokko 2023-01-27
@rdblue thanks for the review. This one is blocked by https://github.com/apache/iceberg/pull/6566